split Bot class
