+++ baby_yoda_bot/constants/text.py
@@ -1,6 +1,4 @@
-TEXT = {
-    "WELCOME": "\nWelcome to Baby Yoda Bot! Here are some magic commands you can use:\n"
-}
+TEXT = {"WELCOME": "\nWelcome to Baby Yoda Bot! Here are some magic commands:\n"}
 
 
 __all__ = ["TEXT"]
